+++ frontend/src/components/Layout.jsx
@@ -1,0 +1,25 @@
+import { Outlet } from 'react-router-dom'
+import Navbar from './Navbar'
+import Sidebar from './Sidebar'
+
+function Layout() {
+  return (
+    <div className="flex h-screen bg-dark-900">
+      {/* Sidebar */}
+      <Sidebar />
+      
+      {/* Main Content */}
+      <div className="flex flex-col flex-1 overflow-hidden">
+        {/* Navbar */}
+        <Navbar />
+        
+        {/* Page Content */}
+        <main className="flex-1 overflow-auto p-6">
+          <Outlet />
+        </main>
+      </div>
+    </div>
+  )
+}
+
+export default Layout